src/requirements.txt
# Core dependencies for OCR ID Card Extractor
google-generativeai>=0.3.0
Pillow>=9.0.0
python-docx>=0.8.11
python-dotenv>=0.19.0

# Additional useful packages
requests>=2.28.0
black
flake8
isort